+++ includes/addons/addon-polls.php
@@ -34,11 +34,11 @@ class UserWallWP_Addon_Polls extends UserWall_WP_Base_Addon {
 
 		// SQL query to create the 'userwall_polls' table
 		$sql_query_polls = "CREATE TABLE IF NOT EXISTS $table_polls (
-            poll_id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+            poll_id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
             question_text TEXT NOT NULL,
             creation_date DATETIME NOT NULL,
             user_id BIGINT UNSIGNED NOT NULL,
-            post_id INT UNSIGNED NOT NULL,
+            post_id BIGINT UNSIGNED NOT NULL,
             INDEX user_id_index (user_id),
             INDEX post_id_index (post_id),
             FOREIGN KEY (user_id) REFERENCES {$wpdb->prefix}users(ID),
@@ -47,10 +47,10 @@ class UserWallWP_Addon_Polls extends UserWall_WP_Base_Addon {
 
 		// SQL query to create the 'userwall_poll_votes' table
 		$sql_query_poll_votes = "CREATE TABLE IF NOT EXISTS $table_poll_votes (
-            vote_id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+            vote_id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
             user_id BIGINT UNSIGNED NOT NULL,
-            poll_id INT UNSIGNED NOT NULL,
-            selected_option INT NOT NULL,
+            poll_id BIGINT UNSIGNED NOT NULL,
+            selected_option BIGINT NOT NULL,
             INDEX user_id_index (user_id),
             INDEX poll_id_index (poll_id),
             FOREIGN KEY (user_id) REFERENCES {$wpdb->prefix}users(ID),
@@ -59,8 +59,8 @@ class UserWallWP_Addon_Polls extends UserWall_WP_Base_Addon {
 
 		// SQL query to create the 'wp_userwall_poll_options' table
 		$sql_query_poll_options = "CREATE TABLE IF NOT EXISTS $table_poll_options (
-            option_id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-            poll_id INT UNSIGNED NOT NULL,
+            option_id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+            poll_id BIGINT UNSIGNED NOT NULL,
             option_text TEXT NOT NULL,
             INDEX poll_id_index (poll_id),
             FOREIGN KEY (poll_id) REFERENCES wp_userwall_polls(poll_id)
